refactor(ui): move CardTemplate to its own file
